requirement.txt
flask
flask-smorest
flask-SQLAlchemy
python-dotenv